pyproject.toml
[project]
name = "adk-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.10.2",
    "google-adk>=1.4.2",
    "litellm>=1.73.0",
    "mcp-neo4j-cypher>=0.2.3",
    "mcp-neo4j-memory>=0.1.5",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
]
